set Youtube folder to system folder
fixing download location, I dont really use this feature so its not optimized yet. I think the cookies might need to be still reset to a set location. 

One issue I remember encountering was that I could only download one file per session. What I was planning for this was to not download the file and just stream it or delete it after. 

Signed-off-by: point.aveugle <83179833+pointave@users.noreply.github.com>