Create video_exporter_cuda.py
Rename to video_exporter.py to export with gpu; it will export in h265 and export faster and quality may be altered. Useful for longer clips. 

Had a branch that did the switch between h264 and h265 seamlessly but it changed way too much so will need to port in next version a good solution.

Signed-off-by: point.aveugle <83179833+pointave@users.noreply.github.com>